Do not store prevalanceDataDate in local storage.
It causes type checking errors on local storage load from the Date being
serialized as a string and could potentially silently inflate numbers in
manual entry mode if the percent increase in cases applies.

prevalance is usually spelled prevalEnce, but this commit does not
change the spelling of that property key to avoid migration issues.